Complete repository layer update with all repository interfaces

+++ src/main/java/org/example/revhire/repository/NotificationRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 @Repository
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
-    List<Notification> findByUserIdOrderByCreatedAtDesc(Integer userId);
+    List<Notification> findByUserIdOrderByCreatedAtDesc(Long userId);
 
-    List<Notification> findByUserIdAndIsReadFalse(Integer userId);
+    List<Notification> findByUserIdAndIsReadFalse(Long userId);
 }
