+++ src/main/java/org/example/revhire/repository/ResumeSkillRepository.java
@@ -1,4 +1,5 @@
 package org.example.revhire.repository;
+
 import org.example.revhire.model.ResumeSkills;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface ResumeSkillRepository extends JpaRepository<ResumeSkills, Integer> {
-    List<ResumeSkills> findByUserId(Integer userId);
+    List<ResumeSkills> findByUserId(Long userId);
 }
